requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
openai==2.21.0
pydantic==2.12.5
python-dotenv==1.2.1
